fixed different issues to get good results

+++ src/ukf.h
@@ -33,7 +33,7 @@ public:
 
   ///* time when the state is true, in us
   long long time_us_;
-  long long previous_t;
+  double previous_t;
 
   ///* Process noise standard deviation longitudinal acceleration in m/s^2
   double std_a_;
